# web_api-main/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "web_api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/automovel": {
      "get": {
        "tags": [
          "Automovel"
        ],
        "summary": "Retorna uma lista de automoveis.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AutomovelModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Automovel"
        ],
        "summary": "Cria uma automovel novo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AutomovelModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/automovel/{id}": {
      "get": {
        "tags": [
          "Automovel"
        ],
        "summary": "Retorna um automovel pelo Id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AutomovelModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Automovel"
        ],
        "summary": "Altera os dados do automovel pelo Id",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AutomovelModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Automovel"
        ],
        "summary": "Deleta o registro de um automovel pelo seu Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/cobertura": {
      "get": {
        "tags": [
          "Cobertura"
        ],
        "summary": "Retorna uma lista de coberturas.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CoberturaModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cobertura"
        ],
        "summary": "Cria uma cobertura nova",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CoberturaModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/cobertura/{id}": {
      "get": {
        "tags": [
          "Cobertura"
        ],
        "summary": "Retorna uma cobertura pelo Id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CoberturaModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cobertura"
        ],
        "summary": "Altera os dados da cobertura pelo Id",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CoberturaModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cobertura"
        ],
        "summary": "Deleta o registro de uma cobertura pelo seu Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/plano": {
      "get": {
        "tags": [
          "Plano"
        ],
        "summary": "Retorna uma lista de planos.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlanosModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plano"
        ],
        "summary": "Cria um plano novo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PlanosModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/plano/{id}": {
      "get": {
        "tags": [
          "Plano"
        ],
        "summary": "Retorna um plano pelo Id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PlanosModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Plano"
        ],
        "summary": "Altera os dados do plano pelo Id",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PlanosModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Plano"
        ],
        "summary": "Deleta o registro de um plano pelo seu Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/seguradora": {
      "get": {
        "tags": [
          "Seguradora"
        ],
        "summary": "Retorna uma lista de seguradoras.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlanoModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Seguradora"
        ],
        "summary": "Cria uma seguradora nova",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PlanoModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/seguradora/{id}": {
      "get": {
        "tags": [
          "Seguradora"
        ],
        "summary": "Retorna uma seguradora pelo Id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PlanoModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Seguradora"
        ],
        "summary": "Altera os dados da seguradora pelo Id",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PlanoModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Seguradora"
        ],
        "summary": "Deleta o registro de uma seguradora pelo seu Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "AutomovelModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "modelo": {
          "type": "string"
        },
        "marca": {
          "type": "string"
        },
        "anoModelo": {
          "type": "string"
        },
        "cor": {
          "type": "string"
        },
        "placa": {
          "type": "string"
        },
        "renavam": {
          "type": "string"
        },
        "numeroMotor": {
          "type": "string"
        },
        "crlv": {
          "type": "string"
        }
      }
    },
...
    "CoberturaModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "idPlano": {
          "format": "int32",
          "type": "integer"
        },
        "descriacao": {
          "type": "string"
        },
        "indenizacao": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
    "PlanoModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "razaoSocial": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "contratoSocial": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        }
      }
    },
    "PlanosModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nomePlano": {
          "type": "string"
        },
        "idSeguradora": {
          "format": "int32",
          "type": "integer"
        },
        "tipoPlano": {
          "$ref": "#/definitions/TipoPlano"
        }
      }
    },
...
    "TipoPlano": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    }
  }
}